+++ android/tv/ui/common/src/main/java/fr/shiningcat/simplehiit/android/tv/ui/common/components/ChoiceDialog.kt
@@ -1,6 +1,5 @@
 package fr.shiningcat.simplehiit.android.tv.ui.common.components
 
-import android.content.res.Configuration
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 import androidx.compose.foundation.Image
@@ -18,14 +17,14 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Devices
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import androidx.tv.material3.MaterialTheme
 import androidx.tv.material3.Surface
 import androidx.tv.material3.SurfaceDefaults
 import androidx.tv.material3.Text
+import fr.shiningcat.simplehiit.android.common.ui.utils.adaptDpToFontScale
+import fr.shiningcat.simplehiit.android.tv.ui.common.previews.PreviewTvScreensNoUi
 import fr.shiningcat.simplehiit.android.tv.ui.common.theme.SimpleHiitTvTheme
 import fr.shiningcat.simplehiit.commonresources.R
 
@@ -71,7 +70,7 @@ fun ChoiceDialog(
                     Image(
                         modifier =
                             Modifier
-                                .size(120.dp)
+                                .size(adaptDpToFontScale(120.dp))
                                 .align(Alignment.CenterHorizontally)
                                 .padding(horizontal = 0.dp, vertical = 24.dp),
                         painter = painterResource(id = image),
@@ -91,20 +90,29 @@ fun ChoiceDialog(
                 ) {
                     if (secondaryButtonLabel.isNotBlank()) {
                         ButtonText(
-                            modifier = Modifier.height(48.dp).weight(1f),
+                            modifier =
+                                Modifier
+                                    .height(adaptDpToFontScale(48.dp))
+                                    .weight(1f),
                             onClick = secondaryAction,
                             label = secondaryButtonLabel,
                         )
                     }
                     if (dismissButtonLabel.isNotBlank()) {
                         ButtonBordered(
-                            modifier = Modifier.height(48.dp).weight(1f),
+                            modifier =
+                                Modifier
+                                    .height(adaptDpToFontScale(48.dp))
+                                    .weight(1f),
                             onClick = dismissAction,
                             label = dismissButtonLabel,
                         )
                     }
                     ButtonFilled(
-                        modifier = Modifier.height(48.dp).weight(1f),
+                        modifier =
+                            Modifier
+                                .height(adaptDpToFontScale(48.dp))
+                                .weight(1f),
                         onClick = primaryAction,
                         label = primaryButtonLabel,
                     )
@@ -115,16 +123,7 @@ fun ChoiceDialog(
 }
 
 // Previews
-@Preview(
-    showSystemUi = true,
-    device = Devices.TV_1080p,
-    uiMode = Configuration.UI_MODE_NIGHT_NO,
-)
-@Preview(
-    showSystemUi = true,
-    device = Devices.TV_1080p,
-    uiMode = Configuration.UI_MODE_NIGHT_YES,
-)
+@PreviewTvScreensNoUi
 @Composable
 private fun ChoiceDialogPreview() {
     SimpleHiitTvTheme {
