+++ android/tv/ui/home/src/main/java/fr/shiningcat/simplehiit/android/tv/ui/home/contents/HomeContentHolder.kt
@@ -9,7 +9,7 @@ import androidx.tv.material3.MaterialTheme
 import androidx.tv.material3.Surface
 import fr.shiningcat.simplehiit.android.tv.ui.common.components.BasicLoading
 import fr.shiningcat.simplehiit.android.tv.ui.common.components.WarningDialog
-import fr.shiningcat.simplehiit.android.tv.ui.common.previews.PreviewTvScreen
+import fr.shiningcat.simplehiit.android.tv.ui.common.previews.PreviewTvScreens
 import fr.shiningcat.simplehiit.android.tv.ui.common.theme.SimpleHiitTvTheme
 import fr.shiningcat.simplehiit.android.tv.ui.home.HomeDialog
 import fr.shiningcat.simplehiit.android.tv.ui.home.HomeViewState
@@ -72,7 +72,7 @@ fun HomeContentHolder(
 
 // Previews
 @ExperimentalTvMaterial3Api
-@PreviewTvScreen
+@PreviewTvScreens
 @Composable
 private fun HomeContentHolderPreviewTV(
     @PreviewParameter(HomeContentHolderPreviewParameterProvider::class) viewState: HomeViewState,
