+++ android/tv/ui/home/src/main/java/fr/shiningcat/simplehiit/android/tv/ui/home/contents/HomeMissingUsersContent.kt
@@ -29,7 +29,7 @@ import androidx.tv.material3.MaterialTheme
 import androidx.tv.material3.Surface
 import androidx.tv.material3.Text
 import fr.shiningcat.simplehiit.android.tv.ui.common.components.ButtonFilled
-import fr.shiningcat.simplehiit.android.tv.ui.common.previews.PreviewTvScreen
+import fr.shiningcat.simplehiit.android.tv.ui.common.previews.PreviewTvScreens
 import fr.shiningcat.simplehiit.android.tv.ui.common.theme.SimpleHiitTvTheme
 import fr.shiningcat.simplehiit.commonresources.R
 import kotlinx.coroutines.delay
@@ -96,7 +96,7 @@ fun HomeMissingUsersContent(navigateToSettings: () -> Unit = {}) {
 
 // Previews
 @ExperimentalTvMaterial3Api
-@PreviewTvScreen
+@PreviewTvScreens
 @Composable
 private fun HomeMissingUsersContentPreviewPhonePortrait() {
     SimpleHiitTvTheme {
