+++ android/tv/ui/session/src/main/java/fr/shiningcat/simplehiit/android/tv/ui/session/SessionScreenTvPreview.kt
@@ -6,7 +6,7 @@ import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import androidx.tv.material3.ExperimentalTvMaterial3Api
 import androidx.tv.material3.MaterialTheme
 import androidx.tv.material3.Surface
-import fr.shiningcat.simplehiit.android.tv.ui.common.previews.PreviewTvScreen
+import fr.shiningcat.simplehiit.android.tv.ui.common.previews.PreviewTvScreens
 import fr.shiningcat.simplehiit.android.tv.ui.common.theme.SimpleHiitTvTheme
 import fr.shiningcat.simplehiit.domain.common.models.Exercise
 import fr.shiningcat.simplehiit.domain.common.models.ExerciseSide
@@ -14,7 +14,7 @@ import fr.shiningcat.simplehiit.domain.common.models.SessionStepDisplay
 
 // Previews
 @ExperimentalTvMaterial3Api
-@PreviewTvScreen
+@PreviewTvScreens
 @Composable
 fun SessionScreenTvPreview(
     @PreviewParameter(SessionScreenPreviewParameterProvider::class) viewState: SessionViewState,
