+++ android/tv/ui/session/src/main/java/fr/shiningcat/simplehiit/android/tv/ui/session/contents/SessionContentHolder.kt
@@ -9,7 +9,7 @@ import androidx.tv.material3.MaterialTheme
 import androidx.tv.material3.Surface
 import fr.shiningcat.simplehiit.android.tv.ui.common.components.BasicLoading
 import fr.shiningcat.simplehiit.android.tv.ui.common.components.ChoiceDialog
-import fr.shiningcat.simplehiit.android.tv.ui.common.previews.PreviewTvScreen
+import fr.shiningcat.simplehiit.android.tv.ui.common.previews.PreviewTvScreens
 import fr.shiningcat.simplehiit.android.tv.ui.common.theme.SimpleHiitTvTheme
 import fr.shiningcat.simplehiit.android.tv.ui.session.CountDown
 import fr.shiningcat.simplehiit.android.tv.ui.session.RunningSessionStepType
@@ -82,7 +82,7 @@ fun SessionContentHolder(
 
 // Previews
 @ExperimentalTvMaterial3Api
-@PreviewTvScreen
+@PreviewTvScreens
 @Composable
 private fun SessionContentHolderPreview(
     @PreviewParameter(SessionContentHolderPreviewParameterProvider::class) viewState: SessionViewState,
