+++ android/tv/ui/settings/src/main/java/fr/shiningcat/simplehiit/android/tv/ui/settings/components/SettingsUsersComponent.kt
@@ -27,7 +27,7 @@ import androidx.tv.material3.Surface
 import androidx.tv.material3.Text
 import fr.shiningcat.simplehiit.android.tv.ui.common.components.ButtonBordered
 import fr.shiningcat.simplehiit.android.tv.ui.common.components.ButtonFilled
-import fr.shiningcat.simplehiit.android.tv.ui.common.previews.PreviewTvScreen
+import fr.shiningcat.simplehiit.android.tv.ui.common.previews.PreviewTvScreens
 import fr.shiningcat.simplehiit.android.tv.ui.common.theme.SimpleHiitTvTheme
 import fr.shiningcat.simplehiit.commonresources.R
 import fr.shiningcat.simplehiit.domain.common.models.User
@@ -98,7 +98,7 @@ fun SettingsUsersComponent(
 }
 
 // Previews
-@PreviewTvScreen
+@PreviewTvScreens
 @Composable
 private fun SettingsUsersComponentPreviewPhonePortrait(
     @PreviewParameter(SettingsUsersComponentPreviewParameterProvider::class) users: List<User>,
